refactor(reminder): add missing field to ReminderGroup type; optimize response type for reminder apis
